Move mock implementations to mock package
