+++ services/job.go
@@ -1,25 +1,9 @@
 package services
 
 import (
-	"bytes"
-	"encoding/json"
-	"errors"
-	"io/ioutil"
-	"net/http"
-	"path"
-	"sync"
-
 	scari "github.com/rafax/scari"
 	uuid "github.com/satori/go.uuid"
 )
-
-type JobStore interface {
-	Put(j scari.Job) error
-	Get(id scari.JobID) (*scari.Job, error)
-	GetAll() ([]scari.Job, error)
-	LeaseOne(scari.LeaseID) (*scari.Job, error)
-	Complete(lid scari.LeaseID, storageID string) (*scari.Job, error)
-}
 
 type JobService interface {
 	New(url string, output scari.OutputType) (*scari.Job, error)
@@ -29,83 +13,15 @@ type JobService interface {
 	Complete(lid scari.LeaseID, storageID string) (*scari.Job, error)
 }
 
-type mapJobStore struct {
-	jobsLock   sync.RWMutex
-	jobs       map[scari.JobID]scari.Job
-	leasedJobs map[scari.LeaseID]scari.JobID
-}
-
-func (m *mapJobStore) Get(id scari.JobID) (*scari.Job, error) {
-	m.jobsLock.RLock()
-	defer m.jobsLock.RUnlock()
-	j, ok := m.jobs[id]
-	if !ok {
-		return nil, errors.New("Job not found")
-	}
-	return &j, nil
-}
-
-func (m *mapJobStore) Complete(leaseID scari.LeaseID, storageID string) (*scari.Job, error) {
-	m.jobsLock.Lock()
-	defer m.jobsLock.Unlock()
-	jid, ok := m.leasedJobs[leaseID]
-	if !ok {
-		return nil, errors.New("Lease not found")
-	}
-	j, ok := m.jobs[jid]
-	if !ok {
-		return nil, errors.New("Job not found")
-	}
-	delete(m.leasedJobs, leaseID)
-	j.Status = scari.Completed
-	j.StorageID = storageID
-	m.jobs[j.ID] = j
-	return &j, nil
-}
-
-func (m *mapJobStore) GetAll() ([]scari.Job, error) {
-	m.jobsLock.RLock()
-	defer m.jobsLock.RUnlock()
-	res := make([]scari.Job, len(m.jobs))
-	i := 0
-	for _, j := range m.jobs {
-		res[i] = j
-		i++
-	}
-	return res, nil
-}
-
-func (m *mapJobStore) Put(j scari.Job) error {
-	m.jobsLock.Lock()
-	defer m.jobsLock.Unlock()
-	m.jobs[j.ID] = j
-	return nil
-}
-
-func (m *mapJobStore) LeaseOne(lid scari.LeaseID) (*scari.Job, error) {
-	m.jobsLock.Lock()
-	defer m.jobsLock.Unlock()
-	for _, j := range m.jobs {
-		if j.Status == scari.Pending {
-			j.Status = scari.Processing
-			m.jobs[j.ID] = j
-			m.leasedJobs[lid] = j.ID
-			// TODO expire locks
-			return &j, nil
-		}
-	}
-	return nil, nil
-}
-
 type jobService struct {
-	store         JobStore
-	storageClient StorageClient
+	store         scari.JobStore
+	storageClient scari.StorageClient
 }
 
-func NewJobService() JobService {
+func NewJobService(store scari.JobStore, storageClient scari.StorageClient) JobService {
 	return &jobService{
-		store:         &mapJobStore{jobs: map[scari.JobID]scari.Job{}, leasedJobs: map[scari.LeaseID]scari.JobID{}},
-		storageClient: &mockStorageClient{client: &http.Client{}}}
+		store:         store,
+		storageClient: storageClient}
 }
 
 func (js *jobService) New(source string, output scari.OutputType) (*scari.Job, error) {
@@ -142,34 +58,4 @@ func (js *jobService) Complete(lid scari.LeaseID, fileName string) (*scari.Job, 
 		return nil, err
 	}
 	return j, err
-}
-
-type StorageClient interface {
-	Register(string) (string, error)
-}
-
-type mockStorageClient struct {
-	client *http.Client
-}
-
-func (msc *mockStorageClient) Register(fileName string) (string, error) {
-	sfr := scari.StaticFileRequest{FileName: path.Base(fileName)}
-	body, err := json.Marshal(sfr)
-	if err != nil {
-		return "", err
-	}
-	resp, err := msc.client.Post("http://scari-666.appspot.com/files", "application/json", bytes.NewReader(body))
-	if err != nil {
-		return "", err
-	}
-	defer resp.Body.Close()
-	rbody, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return "", err
-	}
-	sfresp := new(scari.StaticFileResponse)
-	if err = json.Unmarshal(rbody, sfresp); err != nil {
-		return "", err
-	}
-	return sfresp.Id, nil
 }
